+++ examples/clone/clone.rs
@@ -4,7 +4,7 @@ struct Nil;
 
 // A tuple struct with resources that implements the `Clone` trait
 #[derive(Clone,Show)]
-struct Pair(Box<int>, Box<int>);
+struct Pair(Box<isize>, Box<isize>);
 
 fn main() {
     // Instantiate `Nil`
